#138 - adição da pasta types
