+++ frontend/src/types/ColetaDados.ts
@@ -1,0 +1,8 @@
+export type ColetaDados = {
+  OBindice: string
+  OBqualificacao: string
+  OBavaliacao: string
+  OBpontuacao: string
+  OBcomentario: string
+  OBstatus: string
+}
